feat:hackclub hackathon command
