+++ cogs/Fun.py
@@ -9,7 +9,7 @@ import pyfiglet
 import os
 import requests
 from cogs.usefullTools.dbIntegration import *
-from cogs.usefullTools.info import *
+
 
 def NASA_API_KEY():
 	return os.getenv("NASA_API_KEY")
@@ -368,27 +368,6 @@ class FunCog(commands.Cog):
 			await ctx.send(f'An error occured \n```\n{error}\n```\nPlease check console for traceback, or raise an issue to cyberon')
 			raise error
 			
-	@commands.command(name='hacks')
-	@cooldown(1, 2, BucketType.channel)
-	async def hackathons(self, ctx):
-			url = 'https://hackathons.hackclub.com/api/events/upcoming' 
-			r = requests.get(url)
-			result = r.json()
-			result1 = {}
-			for d in result:
-				result1.update(d)
-				break
-			data = parse_data(result1)
-			await ctx.send(embed = hack_message(data))
-			
-	@hackathons.error
-	async def hackathons_error(self, ctx, error):
-		if isinstance(error, commands.CommandOnCooldown):
-			await ctx.send(error)
-		else:
-			await ctx.send(f'An error occured \n```\n{error}\n```\nPlease check console for traceback, or raise an issue to cyberon')
-			raise error
-
 	@commands.command(name='quotes', aliases=['quote'])
 	@cooldown(1, 2, BucketType.channel)
 	async def quotes(self, ctx):
